Move side effect printing utility
